+++ zkfold-base.cabal
@@ -89,7 +89,6 @@ library
     import: options
     exposed-modules:
       ZkFold.Base.Algebra.Basic.Class
-      ZkFold.Base.Algebra.Basic.Sources
       ZkFold.Base.Algebra.Basic.Field
       ZkFold.Base.Algebra.Basic.Number
       ZkFold.Base.Algebra.Basic.Permutations
@@ -187,6 +186,7 @@ library
       ZkFold.Symbolic.Compiler.ArithmeticCircuit.Instance
       ZkFold.Symbolic.Compiler.ArithmeticCircuit.Internal
       ZkFold.Symbolic.Compiler.ArithmeticCircuit.Map
+      ZkFold.Symbolic.Compiler.ArithmeticCircuit.Sources
       ZkFold.Symbolic.Data.Bool
       ZkFold.Symbolic.Data.ByteString
       ZkFold.Symbolic.Data.Class
